+++ packages/nano_testing/lib/nano_testing.dart
@@ -1,1 +1,2 @@
 export 'src/nano_testing.dart';
+export 'src/fixtures/student_home_fixtures.dart';